Guardar compras en base de datos
git-svn-id: http://jira.uji.es/repos/uji/PAR/uji-par/trunk@13441 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ uji-par-admin/src/main/java/es/uji/apps/par/services/rest/EventosResource.java
@@ -71,7 +71,7 @@ public class EventosResource
     @GET
     @Path("{eventoId}/sesiones/{sesionId}/precios")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getPreciosSesion(@PathParam("eventoId") Integer eventoId, @PathParam("sesionId") Integer sesionId)
+    public Response getPreciosSesion(@PathParam("eventoId") Integer eventoId, @PathParam("sesionId") Long sesionId)
     {
         return Response.ok().entity(new RestResponse(true, sesionesService.getPreciosSesion(sesionId)))
                 .build();
